+++ food_classifier/src/service_ui/pages/customer_page.py
@@ -6,7 +6,7 @@ current_dir = os.path.dirname(os.path.abspath(__file__))
 parent_dir = os.path.abspath(os.path.join(current_dir, '..'))
 sys.path.append(parent_dir)
 
-from utils.customer_processing import CustomerProcessor
+from processors.customer_processing import CustomerProcessor
 
 # Initialize processor
 customer_processor = CustomerProcessor()
